backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
elevenlabs==0.2.26
moviepy==1.0.3
Pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
validators==0.22.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0 